backend/requirements.txt
Django>=4.2,<5.3
djangorestframework>=3.14,<3.16
djangorestframework-simplejwt>=5.3,<6
django-cors-headers>=4.4,<5
django-filter>=24.3
drf-spectacular>=0.27,<0.28
python-decouple>=3.8,<4
dj-database-url>=2.2,<3
psycopg[binary]>=3.2,<4
Pillow>=10.4,<11
celery>=5.4,<6
redis>=5,<6
requests>=2.32,<3
python-dateutil>=2.9,<3
gunicorn>=23,<24
